Exercise Api
Takes in an array of numbers. Each number represents a body part, corresponding body parts can be found in the code comments. Creates a div of 2 different exercises for each of the body parts. Each exercise div consists of the title and the description of the exercise. Image is on the works